Split models/remote_build.py into offloader / receiver siblings
